PEMA-309 Added JWT handling for display name
Acquiring the display name is possible via adding the proper scope (profile), which means that the UI should seamlessly log into oath accounts.

Note that the JWT might be updated or changed, as such to cope with that the token fields should be gotten on call instead of being static.