PEMA-309 Altered log in logic
The process to log in via OAuth2 is now seamless, when the OAuth logging button is pressed and post segue to the login screen, the OAuth2 process will be automatically launched and all the important user data will be set directly from the OAuth request.

TODO: Properly review if moving all the login in logic to a helper would make it cleaner (as to avoid repeating this logic in both the login VM and account type select VM.
TODO: Move the JWT logic to it's own class or library to mirror android and avoid dependency bloat.